My C - Functions, nested loops Task

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -11,10 +11,10 @@
 
 int main(void)
 {
-	unsigned int x = 1;
-	unsigned int y = 2;
-	unsigned int z;
-	unsigned int Sum;
+	long unsigned int x = 1;
+	long unsigned int y = 2;
+	long unsigned int z;
+	long unsigned int Sum;
 
 	printf("1, 2");
